Adds resource type classes
A set of resource types is added in /nova/resources/types.py. Simple
integer-based resources like CPU, RAM, and LocalDisk are included, in
addition to complex resources like NUMATopology, which uses the
primitives in nova.virt.hardware in order to update usage amounts and
calculate whether a requested NUMATopology for a to-be-launched instance
would fit in the NUMA cells advertised and used on a compute node.

A base nova.resources.types.ResourceType class is used as a generic
class interface for any resource, simple or complex, to create
nova.resources.amounts.AmountSpec and UsageSpec objects that can be used
to represent concrete amount and usage records. The idea here is to have
a single way of modeling some resource that is provided by a compute
node, and a single way of comparing a requested amount of something with
the capacity and usage records of the provider of that resource.

The next patch in the series converts the resource tracker to use these
resource amount and usage objects instead of the four different ways
that it currently tracks different types of resources.

Change-Id: I1962469d4d39c7ca2e9632ff8dad1e7f4de9125c
Partially-implements: resource-objects